Add create purchase functionality
Add purchase post call function onto button in IndexMovies.js @rosydiazl @jbeecy
