--- src/components/purchases/CreatePurchase.js
@@ -1,15 +1,0 @@
-// import React, { Component } from 'react'
-// import { withRouter } from 'react-router-dom'
-
-// class CreatePurchase extends Component {
-//   constructor (props) {
-//     super(props)
-//     this.state = {
-//       title: '',
-//       description: '',
-//       price: ''
-//     }
-//   }
-// }
-
-// export default withRouter(CreatePurchase)
